Fixed frame perfect flame color bug
The logic for determining the color of the flame was slightly wrong.
The first frame the fire was turned on, it would always be the same color as the color of player 0 (orange / red).
The next frame, the game would update it to the proper color, only if the next frame the house tile was taking damage as well.
In very rare situations, the tile would not be taking damage the next frame, which would mean that sometimes, the player would notice that the flames on the tile were of the wrong color.
This has now been fixed by making the proper checks within the same frame.