Temporary fix for double scene start
The old system would cause a double scene start. The new system requires manual cleanup of the scenes.

TODO: Find a better solution that has the best of both worlds. (The project already has a few scenes that work with manual cleanup, so this is a must for the future maintainability of the project)