Added fire and destruction sounds to the houses
The house tiles now play a fire ambient sound while they are burning. The fire sound is attenuated based on the distance to the player1 as of now.

The houses now play a destruction sound once they are fully broken. The destruction sound is global and plays across the entire map.