Added support for loading and saving users to disk
Users can now be stored in a JSON file in the hard drive.

When the server is started, the users are loaded from the file. If the file is empty, no users are created.

An admin level GET command was added to manually perform a save of the current users. The user list is updated and stored in the JSON file accordingly.

TODO: Replace the admin level test commands with actual user level interactions and periodic saves or some form of backing system.