Added new buttons to the menu
The main menu now has a placeholder "Social" community button, which will be used to display all of the accounts within the server in the future.

The credits menu now makes use of the Button class so that it automatically has the same behavior as all the other buttons (effects on hover and click).