Created credentials box class
Encapsulated the logic for display and handling a credentials box for account creation, access and deletion, within it's own class.

All the caller has to do is configure the callback function for the ajax request or whatever process that has to be performed (even local process with no internet access that still require account validation).