Added login box
Added partial support for logging in from the client side through ajax requests to the server.

The login box now performs proper GET requests and properly handles finding the correct user and logging in when the credentials given are valid.

TODO: Start displaying the user's profile in the account button once the user is logged in.