Added square buttons next to user names
The users list now displays a squre button next to the users' names to signify that the current line holds an user's name and that it can be interacted with.

This way, users with invisible characters will still be spottable on the list.

This also acts as a form of decoration.