Added working forum chat
The clientside of the chat is now almost fully functional.

It will display messages sent by the local user, and send them to the server.

The user will update the messages list by getting the messages stored in the server.

TODO: Add auto scroll when a new message is received and add proper user ID to sent messages.