Banned some special characters from username
Due to how HTML files are displayed and how the forum chat is currently generated (using innerHTML), certain special characters have been temporarily banned from being used in the username.

This fix does not work for chat messages. Users can currently insert HTML elements, images, videos, etc... (some content does not work properly due to the fact that we're currently forced to use a constantly reloading REST API instead of sockets, so videos will always be restarted (their element is deleted and regenerated)).

TODO: Actually fix the real issue and make use of textContent instead of innerHTML to add the username and message text to the text box. (Probably generate a temporary DOM element and append it to the chat div after modifying the actual textual content)

Optional: keep embeddable HTML in the chat but actually implement it properly in the future.