Fixed chat HTML injection
The getMessageString function now replaces the special characters used by the user with their corresponding HTML sequences to prevent the user from injecting HTML code into the messages.

This might get reverted in the future if a better alternative is designed (for example, adding an HTML mode option for messages, as well as not storing the user's password in the localUser variable and forcing the user to rewrite it every single time it is required to prevent HTML messages with buttons from being usable to steal private user data).