Changed Credentials box to use enum for form type
The form type is now determined through an enum value to prevent needing tons of variables to determine the form type to display.

The children to seek within the HTML file are now determined through an array of children, which is configured by the same switch that determines the form file to load based on the FOMR_TYPE enum value.