Created system to manage user connection status
The clients will now notify the server every few seconds about their connectivity status (only if they are logged into an account). That way, the server will be able to determine which accounts to "keep alive".

The users list can now make use of this system to display the status of the users, by asking the server which users are currently connected. Users that are connected appear with a green square by the side of their username.