Added backbone system for scaling the scene
A backbone system for scaling the Phaser scene has been created.

It does not fully rely on Phaser's default full screen system, as the purpose of this system is to allow the user to also have the game rendered at their native resolution, or whatever chosen resolution the user prefers, which requires a manual implementation instead of scaling the image to FIT, which also causes stretching.

Another strong point of this implementation is the fact that full screen and scaling methods offered by the ScaleManager's full screen system do not work in some devices such as iOS (not the main target audience or even a primary worry for this project, but still, this is another strong point for using and maintaining a custom implementation).