Made visual adjustments to menus
The menus now have a few visual adjustments that were requested by some of the users.

The forum chat's message input box now has a background to make it more intuitive.

The main menu now discards the quit button and has divided the buttons in 2 columns depending on the type of action they perform (game related or user related)