Added support to send messages with the enter key
Pressing the Enter key in the chat text input box now sends the message as any sane user would expect of any reasonable messaging program.

Gone are the days of pressing the "send" button with the mouse or using TAB and SHIFT TAB for quick DOM element context switching hackery.