Fixed backslash exploit in messages on serverside
There was a vulnerability where messages posted in the forum chat that included a backslash (\) would cause the message database to break, as the missmatched backslashes would break the JSON string format.

This has been fixed by adding backslash filtering, as well as quotation mark filtering to the database.

Important note: the backslash filtering has to be done first so as to prevent replacing backslashes added by the filter system itself.