Revamped forum chat command system
The command system now evaluates the text written by the user as JS function, which means that greater flexibility is supported both for users writing commands and for programmers adding more commands to the chat.

Known issues: Command chaining is still in the works. The commands run properly, but the moment a command is chained, the evaluation allows for any JS code to be executed instead of limiting to /command commands. Also, the evaluation result of semicolon separated instructions is that of the last command in the command chain.

Command results can be chained properly by using the plus operator, as /command commands return a string with the resulting HTML to be displayed in the chat.