Changed command system to fix HTML issues
The command msg field was pasted as raw text without any form of processing inside of the HTML, which meant that users could still inject HTML without using the HTML command (wether be it on purpose or by writing special characters such as < and > which would lead to tags being inserted accidentally in the code).

Now, the msg is processed within the command function itself by using the stringReplaceHTMLSymbols() function.

The HTML command has been modified to replace all mentions of localUser.user.password with trollUser, which contains a warning message for people trying to inject HTML to steal passwords.

Example code for demonstration purposes:
/command.html("<button onclick='document.getElementById(&quot;message-input&quot;).value = &quot;Mi contraseña es: &quot; + localUser.user.password; document.getElementById(&quot;send-message&quot;).click();'>No me pulses!</button>")