Added username character validation to server side
The server now checks if incoming requests to create users contain illegal characters or not in the username.

This way, the user is prevented from manually creating an username that could be problematic through an ajax request made from the console or by modifying the login system.

In the past, only the credentials box class on the clientside was in charge of performing illegal character checks for username validation, which is not safe.

TODO: As of now, users are still allowed to create passwords that contain illegal characters. Fix that ASAP.