Added backspace to illegal character check
Users were still allowed to input backspace characters, completely breaking strings in the stored JSON file on the server side. This was fixed with the last character filtering commit. but it would be ideal to prevent users from inputting this character in the first place so as to maintain consistency with the illegal character filtering policy that we have had up until now.

The backspace character has now been  made illegal by adding it to the illegal characters lists in both the server and the client side.