Added Menu for matchmaking and server browser
The online multiplayer submenu has been divided in 2 new menus.

One for automatic matchmaking. The user will press the button and the server will find a suitable lobby to join the user.

The other button will display a server browser, which contains a list of all lobbies within the server.

Currently only the UI is implemented and no internal logic exists on the clientisde.