Created a new class for creation of paged menus
The users list is what this project calls a "paged menu". The tutorial screen and lobbies screen could also use a paged menu.

A custom class has been created to facilitate this process.

TODO: Fix bug where modifying element count of the paged menu to a lower count while the user is in a further page causes index out of bounds reads.

Modify the class so that it automatically resets the current page to a page that is within bounds in the update function.