WIP Warning popup menu
A work in progress warning popup menu.

This menu's purpose is to notify the users that they are about to begin playing multiplayer while they are not logged into any account.

This might get scrapped in favour of either allowing players to play as anonymous or simply taking players to the account menu instead.

This could also be modified later to display an error message when something unexpected happens.