Added support for matchmaking and lobby creation
The client can now fully communicate with the server to do automatic matchmaking.

All of the things that happen under the hood have been covered.

When a player starts matchmaking, they are put in a lobby which is meant to display the connected users and start a match once both players in the lobby are ready to begin.

TODO: Make it so that lobbies can now display user info and begin a match when both players enter the ready state. Also, in the serverside, make it so that lobbies have a state which determines wether a match has started in it or not.