Added support for robust disconnection handling
When a client is disconnected, wether it is manually or through an error (connection is lost or whatever), the disconnection is handled properly and the clients receive a message with the updated status of the game.

An example of this is how now, in a lobby, when a player disconnects, the lobby automatically notifies all remaining players and the text displaying the names as well as the controller character are all updated automatically.