Added support for starting a multiplayer match
The lobby scene can now start a multiplayer match when all players are in a ready state.

This is a WIP, the multiplayer mode does not work yet (still have to update client info with sendData function), but at least the map can now be loaded while keeping players connected to the lobby within the server.

TODO:
-Disconnect players when the game ends and when a connection error happens on one of the clients.
-Add proper multiplayer support (as mentioned before, data replication between clients)