Refactored player and connection data
The player data has been refactored to make use of doubles for the vector position, as well as a double for the rotation angle instead of the wrong type vector.

The lobby scene now makes use of a simpler function to send the data that must be shared with the other remote clients in the lobby.