Added time field to player data
The player data in both the client and the server now contains a time field which contains an integer number that contains the currently elapsed time during the match.

TODO: The objective is for the lobby to dispatch to the remote clients the smallest time value between the 2 clients so that they can update their times accordingly in case of desync.