Refactored game time system and decoupled it
The game timer system has now been refactored so that it is its own self contained object.

The game timer logic is now independent from the UI scene, which makes it easier to access the game time (which will be useful for networking and more complex gameplay mechanics), as it is now its own global object with its own functionality.