Changed sound fix
The sound fix that had been previously created would prevent the sounds from the game from continuing to play in the main menu after finishing a match by disabling and re-enabling sounds when loading the main menu.

This fix would lead to a problem where the sound of buttons from menus that led back to the main menu would be cut off early.

This new fix addresses the issue correctly by disabling the sounds when we return to the main menu.