Created a base class for gameplay logic
The logic from the test maps has been moved to an identical class which implements the game world logic. This class will be used to implement the single player and multiplayer maps. (Either through inheritance or by having a switch that changes the logic used depending on the multiplayer type configuration.)