Reworked lobby player identification system
The player class on the server side now stores 2 IDs: the Account User ID and the Lobby Player ID.

The user ID is the ID of the account that is being used by said player. This ID will be used for name identification as well as storing progress and stats on the server once the match is finished.

The lobby player ID has now changed its purpose. Instead of storing the User ID, it now stores an ID that is unique for each player within the lobby, which helps with identifying each individual player in an N player lobby. (Even tho this game has only 2 players at most as of now).

The player ID within the current lobby, as well as the current lobby ID are both transmitted to each player so that they know how to identify themselves within the player data array that is received by each client during runtime.