Reworked game logic to make use of player array
The players are now stored as an array instead of 2 variables.

The purpose of this change is to make it easier to implement N player lobbies in the future.

For now, the UI and game over screens all work with 2 players in mind, which means that this change is pretty pointless as of now.