Modified the way the isShooting system works
The dragon's isShooting mechanic has been modified to slightly facilitate multiplayer support in the future.

currently known issue: this.isShooting = false at the end of the update function makes it impossible for the isShooting to be ever be true when sent through the network. This is done on purpose for current testing purposes and will be fixed once a better system is implemented.