Fixed lobbies due to desync not closing issue
When all players within a lobby would finish the game, they would disconnect. This meant that, due to the multithreaded and async nature of SpringBoot, the server would not properly register the correct number of users within the lobby, leading to lobbies that would not close themselves.

Instead of making an unsafe close lobby command, it has been decided that the lobby will now make use of a synchronized call block on disconnect to prevent lobbies from breaking.