Increased the speed at which packets are sent
Since the sync problem was fixed on the server side (it used to freeze one of the clients sometimes), the speed at which packets can be sent has been increased to a more reasonable value.

This value should be tweaked within capacity to prevent wasting resources and using as many as are truly required to obtain a good gameplay experience.