Fixed issues with game freezing and crashing
Someone had apparently made use of variables from the dragon class in the UI class and flame group file without actually creating the variable in the constructor of the dragon class.

The property was actually created for the first time in the flame group file.

This meant that, due to a miracle, things have worked up until now.

This issue has now been resolved by actually creating the property within the dragon class.