Added Settings menu with volume options
There is now a settings menu that has a set of options to allow the user to regulate the volume of the game.

The game currently allows the user to regulate the volume of in game effects, menu buttons and music (even tho there is no music as of now in the game, the system to support it is already implemented).