Added basic setup for ingame settings menu button
The UI scene now has a settings button that is displayed in game.

When the player hovers the button, it will start rotating.

WIP. The button does not do anything as of now. It is meant to open a pause menu / settings menu (the game will not pause due to its multiplayer nature)