Added some groundwork for in game menu
The settings menu now has an init method that takes a data object that contains a variable that informs the settings menu about whether it was opened from the main menu or from the gameplay scene.

That way, the settings menu will be capable of determining in what way to display the settings menu and how to exit the menu once the user is done using it (the scene unloading and loading process depends on this information to work properly. In the future, all scenes should be reworked to have an init method and take as a parameter the previous scene's key to facilitate exiting scenes without having to refactor everything each time a change in structure happens).