Fixed issue where receiving wrong packets in lobby
Receiving wrong packets in the lobby scene was not handled properly, leading to error messages and other silent JS errors. This has now been fixed.

TODO: Fix UI text resetting to a wrong state when a wrong packet is received (this does get fixed automatically when the next packet received is correct tho)