Updated game over menu to display self user
The game over screen now displays a small text in parenthesis next to the name of the user that corresponds to the owning client. That way, if 2 players are anonymous or have similar names, it will be easier to identify which player from the game over score list is the player that corresponds to the client that is playing.

Also, added support for anonymous users to have the anonymous username displayed.

Known issues: The self annotation in spanish "(Tú)" contains a tilde, which does not render with the current font.