Changed font in game over screen
The game over screen now displays certain texts with different fonts.

This has been done to make the menu capable of displaying certain special characters that were not present in the previously chosen fonts, as well as making the menu look nicer with the better looking fonts that have been chosen.