Added in game menus (pause and settings menus)
The game now supports displaying a "pause" menu from which the player can choose to resume the game, display the settings and quit.

The settings menu has been modified slightly so as to support being displayed by being accessed both from the main menu and from the in game pause menu.

TODO: handle closing the menu when the time runs out.