Revamped and fixed world sharing system
Reworked the system for sharing the world data and generating the map. When an user minimized the game, they would miss the only time a message was sent to notify the remote clients with the generated world info. This has now been fixed by making the client send the petition to obtain the world data continously, and the host will keep sending the data until all clients have been fed the world data. If a client takes a longer time to join, they will be able to ask for the world data and player status as soon as they are ready.

TODO: Fix issue where users minimizing take longer to end the game, thus having an out of sync game clock, which can cause errors when detecting valid disconnections at the end of the game.