Fixed UI wrong player health displayed bug
The UI used to display the health of each player on the wrong side of the screen depending on wether they were the lobby leader or the remote client.

This happened because of the usage of players[index] as a way to determine which player's data to display, because players[0] is always the lobby leader, which was not accounted for.

This has now been fixed, and the UI properly displays all information.