Fixed player ordering
The players are now ordered by insertion order when the player list is retrieved.

This property is extremely useful and has been missed during development.

The reason why this property was assumed to exist at first and then seemingly lost (which spawned the requirement for using the isLobbyleader function and such, as well as generating some lobby bugs), was that the server implementation moved away from using a simple array to using a concurrent map, which does not guarantee the insertion order to be the same order in which objects are retrieved from the map when transformed into a list.

This has now been fixed by making the server order the list when returning it.