Removed AJAX petition from lobby
The lobby no longer uses an ajax petition to obtain the user names.

This method was dangerous and sometimes lead to a phaser "cut" error, which was caused due to the fact that the ajax petition sometimes would not be fast enough to modify the "ready" text phaser element, leading to errors when this element had already been unloaded from memory when both players were ready and the GameMap scene had been loaded.

This has now been fixed, as well as making the system slimmer and more robust.