Added known issue comment for future reference
Dumbest commit ever to be made. It simply exists so that we don't forget this certain detail about Phaser's implementation.

Literally the only reason this is being added into a commit is that we're packing tomorrow and it would be ideal for this information to be in the repo and not forgotten in someone's personal computer.